+++ R/prepare_report.R
@@ -10,11 +10,13 @@
 #' @param local.timezone Which time zone to use for the date of the report.
 #' @export
 #' @examples
-#' list.prices <- prepare_list_prices(slug = "bitcoin", start.date = "2021-01-01")
+#' list.prices <- list_prices_example
 #' all.data <- format_shakepay(data_shakepay)
 #' formatted.ACB <- format_ACB(all.data, verbose = FALSE)
-#' x <- prepare_report(formatted.ACB, list.prices = list.prices)
-#' x$proceeds
+#' if (!is.null(list.prices)) {
+#'   x <- prepare_report(formatted.ACB, list.prices = list.prices)
+#'   x$proceeds
+#' }
 prepare_report <- function(formatted.ACB,
                            list.prices = NULL,
                            tax.year = "all",